fix: return back typography utils
