+++ packages/picasso/src/Typography/utils/get-typography-class-name/index.ts
@@ -1,0 +1,1 @@
+export { default } from './get-typography-class-name'
